separate package for auth, fix session migration
