+++ src/web/backend/auth/auth.go
@@ -1,4 +1,4 @@
-package backend
+package auth
 
 import (
 	"net/http"
