+++ src/web/backend/auth/middleware.go
@@ -1,4 +1,4 @@
-package backend
+package auth
 
 import (
 	"fmt"
